examples/p2p-coordinated-pipeline/requirements.txt
libp2p==0.6.0
trio==0.33.0
pydantic==2.12.5
python-dotenv>=1.0.0
